Pre: move Either to mentat_core::util.
